Refactor: Improve professionalism of Dotmini MCX application.
This commit introduces several changes to enhance the professionalism of the codebase:

- Removed the redundant `ml.py` file, which was a duplicate of `main.py`.
- Refactored `main.py` by separating major components into their own modules:
    - `ThemeManager` into `theme_manager.py`
    - `LicenseManager` and `LicenseDialog` into `license_manager.py`
    - `FileSelectionWidget` and `FolderSelectionWidget` into `file_widgets.py`
    - `BrightnessControl` into `brightness_control.py`
    - `SettingsDialog` into `settings_dialog.py`
    - `ClassificationThread` and `load_model_safely` into `classification.py`
- Updated `README.md` to include an English section with project details, usage instructions, and a note about the demo license key.
- Addressed the hardcoded demo license key by:
    - Adding a comment in `license_manager.py` to clarify it's for demo purposes.
    - Removing the UI tip that displayed the demo key.
- Added comprehensive docstrings (module, class, and method-level) to all refactored modules and main application classes to improve code understanding and maintainability.
- Minor logical improvements and bug fixes were made during the refactoring and docstring addition process.
