+++ main.py
@@ -1,3 +1,9 @@
+"""Main application module for Dotmini MCX.
+
+This module initializes and runs the Dotmini MCX application,
+setting up the main window, theme management, license verification,
+and coordinating various UI components and functionalities.
+"""
 import os
 import sys
 import hashlib
@@ -24,16 +30,9 @@ from PyQt6.QtGui import (QIcon, QPixmap, QFont, QPalette, QColor, QCursor, QFont
 # For image classification
 import numpy as np
 import tensorflow as tf
-from tensorflow.keras.preprocessing import image
+from tensorflow.keras.preprocessing import image # Retained as it's a common Keras import
 from PIL import Image
 import glob
-
-# For detecting OS theme
-try:
-    from darkdetect import isDark, theme, listener
-    DARKDETECT_AVAILABLE = True
-except ImportError:
-    DARKDETECT_AVAILABLE = False
 
 # Performance optimization - enable TF GPU memory growth
 try:
@@ -54,1272 +53,55 @@ CONFIG_PATH = os.path.join(QStandardPaths.writableLocation(QStandardPaths.Standa
 # Ensure config directory exists
 os.makedirs(CONFIG_PATH, exist_ok=True)
 
-class ThemeManager:
-    def __init__(self, app, main_window=None):
-        self.app = app
-        self.main_window = main_window
-        self.current_theme = "dark"  # Default theme
-        self.brightness = 100  # Default brightness (percentage)
-        self.settings = QSettings(ORGANIZATION_NAME, APP_NAME)
-        
-        # Load saved preferences
-        self.load_settings()
-        
-        # Set up theme detection
-        if DARKDETECT_AVAILABLE:
-            self.detect_system_theme()
-            listener(self.on_system_theme_change)
-        
-        # Apply initial theme
-        self.apply_theme()
-    
-    def load_settings(self):
-        theme_mode = self.settings.value("theme/mode", None)
-        if theme_mode:
-            self.current_theme = theme_mode
-        
-        saved_brightness = self.settings.value("theme/brightness", None)
-        if saved_brightness is not None:
-            try:
-                self.brightness = int(saved_brightness)
-            except (ValueError, TypeError):
-                self.brightness = 100
-    
-    def save_settings(self):
-        self.settings.setValue("theme/mode", self.current_theme)
-        self.settings.setValue("theme/brightness", self.brightness)
-    
-    def detect_system_theme(self):
-        if DARKDETECT_AVAILABLE:
-            self.current_theme = "dark" if isDark() else "light"
-            self.save_settings()
-    
-    def on_system_theme_change(self):
-        if DARKDETECT_AVAILABLE:
-            new_theme = "dark" if isDark() else "light"
-            if new_theme != self.current_theme:
-                self.current_theme = new_theme
-                self.save_settings()
-                self.apply_theme()
-    
-    def set_brightness(self, value):
-        self.brightness = value
-        self.save_settings()
-        self.apply_theme()
-    
-    def toggle_theme(self):
-        self.current_theme = "light" if self.current_theme == "dark" else "dark"
-        self.save_settings()
-        self.apply_theme()
-    
-    def apply_theme(self):
-        # Generate CSS based on theme and brightness
-        if self.current_theme == "dark":
-            # Calculate adjusted colors based on brightness
-            brightness_factor = self.brightness / 100.0
-            
-            # Background colors
-            bg_color = self._adjust_color("#1e1e1e", brightness_factor)
-            card_bg_color = self._adjust_color("#2d2d2d", brightness_factor)
-            input_bg_color = self._adjust_color("#333333", brightness_factor)
-            
-            # Text colors
-            text_color = self._adjust_color("#ffffff", brightness_factor)
-            secondary_text = self._adjust_color("#bbbbbb", brightness_factor)
-            
-            # Accent colors
-            accent_color = self._adjust_color("#0071e3", brightness_factor)
-            accent_hover = self._adjust_color("#0077ED", brightness_factor)
-            accent_pressed = self._adjust_color("#005BBB", brightness_factor)
-            border_color = self._adjust_color("#3d3d3d", brightness_factor)
-            
-            # Generate stylesheet
-            stylesheet = f"""
-                QMainWindow, QDialog, QWidget#splashContent {{
-                    background-color: {bg_color};
-                }}
-                QLabel {{
-                    color: {text_color};
-                    font-family: 'SF Pro Display', 'Helvetica Neue', Arial, sans-serif;
-                }}
-                QLineEdit, QComboBox, QListWidget {{
-                    padding: 8px;
-                    border-radius: 6px;
-                    background-color: {input_bg_color};
-                    color: {text_color};
-                    border: 1px solid {border_color};
-                    font-size: 13px;
-                }}
-                QPushButton {{
-                    padding: 10px 15px;
-                    border-radius: 6px;
-                    background-color: {accent_color};
-                    color: {text_color};
-                    border: none;
-                    font-weight: bold;
-                    font-size: 13px;
-                }}
-                QPushButton:hover {{
-                    background-color: {accent_hover};
-                }}
-                QPushButton:pressed {{
-                    background-color: {accent_pressed};
-                }}
-                QPushButton:disabled {{
-                    background-color: #555555;
-                    color: #888888;
-                }}
-                QPushButton#secondaryButton {{
-                    background-color: {input_bg_color};
-                }}
-                QPushButton#secondaryButton:hover {{
-                    background-color: #444444;
-                }}
-                QToolBar {{
-                    background-color: {card_bg_color};
-                    border: none;
-                    spacing: 10px;
-                    padding: 5px;
-                }}
-                QStatusBar {{
-                    background-color: {card_bg_color};
-                    color: {secondary_text};
-                    border-top: 1px solid {border_color};
-                }}
-                QProgressBar {{
-                    border: none;
-                    border-radius: 6px;
-                    background-color: {input_bg_color};
-                    text-align: center;
-                    color: {text_color};
-                    font-weight: bold;
-                }}
-                QProgressBar::chunk {{
-                    background-color: {accent_color};
-                    border-radius: 6px;
-                }}
-                QListWidget {{
-                    background-color: {input_bg_color};
-                    border-radius: 6px;
-                    color: {text_color};
-                    padding: 5px;
-                }}
-                QListWidget::item {{
-                    border-radius: 4px;
-                    padding: 8px;
-                    margin: 2px 0px;
-                }}
-                QListWidget::item:selected {{
-                    background-color: {accent_color};
-                }}
-                QScrollArea {{
-                    border: none;
-                    background-color: {bg_color};
-                }}
-                QFrame#card {{
-                    background-color: {card_bg_color};
-                    border-radius: 10px;
-                    border: 1px solid {border_color};
-                }}
-                QFrame#resultItem {{
-                    background-color: {card_bg_color};
-                    border-radius: 6px;
-                    border: 1px solid {border_color};
-                    padding: 10px;
-                    margin: 5px;
-                }}
-                QSlider::groove:horizontal {{
-                    height: 8px;
-                    background: #555555;
-                    border-radius: 4px;
-                }}
-                QSlider::handle:horizontal {{
-                    background: {accent_color};
-                    border: none;
-                    width: 18px;
-                    margin: -5px 0;
-                    border-radius: 9px;
-                }}
-                QSlider::add-page:horizontal {{
-                    background: #555555;
-                    border-radius: 4px;
-                }}
-                QSlider::sub-page:horizontal {{
-                    background: {accent_color};
-                    border-radius: 4px;
-                }}
-                QCheckBox {{
-                    color: {text_color};
-                }}
-                QCheckBox::indicator {{
-                    width: 18px;
-                    height: 18px;
-                    border-radius: 4px;
-                    border: 1px solid {accent_color};
-                }}
-                QCheckBox::indicator:checked {{
-                    background-color: {accent_color};
-                }}
-                QMessageBox {{
-                    background-color: {bg_color};
-                }}
-                QMessageBox QLabel {{
-                    color: {text_color};
-                }}
-                QMenu {{
-                    background-color: {card_bg_color};
-                    color: {text_color};
-                    border: 1px solid {border_color};
-                    border-radius: 6px;
-                    padding: 5px;
-                }}
-                QMenu::item {{
-                    padding: 6px 25px 6px 20px;
-                    border-radius: 4px;
-                }}
-                QMenu::item:selected {{
-                    background-color: {accent_color};
-                }}
-                QMenuBar {{
-                    background-color: {bg_color};
-                    color: {text_color};
-                }}
-                QMenuBar::item {{
-                    padding: 5px 10px;
-                    border-radius: 4px;
-                }}
-                QMenuBar::item:selected {{
-                    background-color: {accent_color};
-                }}
-                QToolTip {{
-                    background-color: {card_bg_color};
-                    color: {text_color};
-                    border: 1px solid {border_color};
-                    border-radius: 4px;
-                    padding: 5px;
-                }}
-            """
-        else:  # Light theme
-            brightness_factor = self.brightness / 100.0
-            
-            # Calculate adjusted colors
-            bg_color = self._adjust_color("#f5f5f7", brightness_factor)
-            card_bg_color = self._adjust_color("#ffffff", brightness_factor)
-            input_bg_color = self._adjust_color("#e7e7e7", brightness_factor)
-            text_color = self._adjust_color("#121212", brightness_factor)
-            secondary_text = self._adjust_color("#555555", brightness_factor)
-            accent_color = self._adjust_color("#0071e3", brightness_factor)
-            accent_hover = self._adjust_color("#0077ED", brightness_factor)
-            accent_pressed = self._adjust_color("#005BBB", brightness_factor)
-            border_color = self._adjust_color("#d1d1d1", brightness_factor)
-            
-            # Generate stylesheet
-            stylesheet = f"""
-                QMainWindow, QDialog, QWidget#splashContent {{
-                    background-color: {bg_color};
-                }}
-                QLabel {{
-                    color: {text_color};
-                    font-family: 'SF Pro Display', 'Helvetica Neue', Arial, sans-serif;
-                }}
-                QLineEdit, QComboBox, QListWidget {{
-                    padding: 8px;
-                    border-radius: 6px;
-                    background-color: {input_bg_color};
-                    color: {text_color};
-                    border: 1px solid {border_color};
-                    font-size: 13px;
-                }}
-                QPushButton {{
-                    padding: 10px 15px;
-                    border-radius: 6px;
-                    background-color: {accent_color};
-                    color: white;
-                    border: none;
-                    font-weight: bold;
-                    font-size: 13px;
-                }}
-                QPushButton:hover {{
-                    background-color: {accent_hover};
-                }}
-                QPushButton:pressed {{
-                    background-color: {accent_pressed};
-                }}
-                QPushButton:disabled {{
-                    background-color: #c0c0c0;
-                    color: #888888;
-                }}
-                QPushButton#secondaryButton {{
-                    background-color: {input_bg_color};
-                    color: {text_color};
-                }}
-                QPushButton#secondaryButton:hover {{
-                    background-color: #d5d5d5;
-                }}
-                QToolBar {{
-                    background-color: {card_bg_color};
-                    border: none;
-                    spacing: 10px;
-                    padding: 5px;
-                }}
-                QStatusBar {{
-                    background-color: {card_bg_color};
-                    color: {secondary_text};
-                    border-top: 1px solid {border_color};
-                }}
-                QProgressBar {{
-                    border: none;
-                    border-radius: 6px;
-                    background-color: {input_bg_color};
-                    text-align: center;
-                    color: {text_color};
-                    font-weight: bold;
-                }}
-                QProgressBar::chunk {{
-                    background-color: {accent_color};
-                    border-radius: 6px;
-                }}
-                QListWidget {{
-                    background-color: {input_bg_color};
-                    border-radius: 6px;
-                    color: {text_color};
-                    padding: 5px;
-                }}
-                QListWidget::item {{
-                    border-radius: 4px;
-                    padding: 8px;
-                    margin: 2px 0px;
-                }}
-                QListWidget::item:selected {{
-                    background-color: {accent_color};
-                    color: white;
-                }}
-                QScrollArea {{
-                    border: none;
-                    background-color: {bg_color};
-                }}
-                QFrame#card {{
-                    background-color: {card_bg_color};
-                    border-radius: 10px;
-                    border: 1px solid {border_color};
-                }}
-                QFrame#resultItem {{
-                    background-color: {card_bg_color};
-                    border-radius: 6px;
-                    border: 1px solid {border_color};
-                    padding: 10px;
-                    margin: 5px;
-                }}
-                QSlider::groove:horizontal {{
-                    height: 8px;
-                    background: #c0c0c0;
-                    border-radius: 4px;
-                }}
-                QSlider::handle:horizontal {{
-                    background: {accent_color};
-                    border: none;
-                    width: 18px;
-                    margin: -5px 0;
-                    border-radius: 9px;
-                }}
-                QSlider::add-page:horizontal {{
-                    background: #c0c0c0;
-                    border-radius: 4px;
-                }}
-                QSlider::sub-page:horizontal {{
-                    background: {accent_color};
-                    border-radius: 4px;
-                }}
-                QCheckBox {{
-                    color: {text_color};
-                }}
-                QCheckBox::indicator {{
-                    width: 18px;
-                    height: 18px;
-                    border-radius: 4px;
-                    border: 1px solid {accent_color};
-                }}
-                QCheckBox::indicator:checked {{
-                    background-color: {accent_color};
-                }}
-                QMessageBox {{
-                    background-color: {bg_color};
-                }}
-                QMessageBox QLabel {{
-                    color: {text_color};
-                }}
-                QMenu {{
-                    background-color: {card_bg_color};
-                    color: {text_color};
-                    border: 1px solid {border_color};
-                    border-radius: 6px;
-                    padding: 5px;
-                }}
-                QMenu::item {{
-                    padding: 6px 25px 6px 20px;
-                    border-radius: 4px;
-                }}
-                QMenu::item:selected {{
-                    background-color: {accent_color};
-                    color: white;
-                }}
-                QMenuBar {{
-                    background-color: {bg_color};
-                    color: {text_color};
-                }}
-                QMenuBar::item {{
-                    padding: 5px 10px;
-                    border-radius: 4px;
-                }}
-                QMenuBar::item:selected {{
-                    background-color: {accent_color};
-                    color: white;
-                }}
-                QToolTip {{
-                    background-color: {card_bg_color};
-                    color: {text_color};
-                    border: 1px solid {border_color};
-                    border-radius: 4px;
-                    padding: 5px;
-                }}
-            """
-        
-        self.app.setStyleSheet(stylesheet)
-        
-        # Update main window UI elements if needed
-        if self.main_window:
-            self.main_window.update_theme_ui()
-    
-    def _adjust_color(self, hex_color, brightness_factor):
-        """Adjust color brightness based on the brightness factor"""
-        # Convert hex to RGB
-        hex_color = hex_color.lstrip('#')
-        r, g, b = tuple(int(hex_color[i:i+2], 16) for i in (0, 2, 4))
-        
-        if brightness_factor < 1.0:  # Darkening
-            r = int(r * brightness_factor)
-            g = int(g * brightness_factor)
-            b = int(b * brightness_factor)
-        elif brightness_factor > 1.0:  # Lightening
-            # Calculate how much room we have to brighten
-            r = int(min(r + (255 - r) * (brightness_factor - 1), 255))
-            g = int(min(g + (255 - g) * (brightness_factor - 1), 255))
-            b = int(min(b + (255 - b) * (brightness_factor - 1), 255))
-        
-        # Return as hex
-        return f"#{r:02x}{g:02x}{b:02x}"
+# Import ThemeManager
+from theme_manager import ThemeManager
 
+# Import LicenseManager and LicenseDialog
+from license_manager import LicenseManager, LicenseDialog
 
-class LicenseManager:
-    def __init__(self):
-        self.settings = QSettings(ORGANIZATION_NAME, APP_NAME)
-        self.correct_key = "D1QE80fxUUVcNs4VAAOvNNkJvHHy0dWM"
-    
-    def is_valid_license(self):
-        """Check if a valid license key is already stored"""
-        saved_key = self.settings.value("license/key", None)
-        saved_hash = self.settings.value("license/hash", None)
-        
-        if saved_key and saved_hash:
-            # Verify hash to prevent tampering
-            computed_hash = self._hash_key(saved_key)
-            if computed_hash == saved_hash and saved_key == self.correct_key:
-                return True
-        
-        return False
-    
-    def save_license(self, key):
-        """Save a valid license key"""
-        if key == self.correct_key:
-            key_hash = self._hash_key(key)
-            self.settings.setValue("license/key", key)
-            self.settings.setValue("license/hash", key_hash)
-            return True
-        return False
-    
-    def _hash_key(self, key):
-        """Generate a hash of the license key with a salt"""
-        salt = b"DotminiMCX_salt"  # Use a consistent salt
-        key_bytes = key.encode('utf-8')
-        return hashlib.sha256(key_bytes + salt).hexdigest()
+# Import classification components
+from classification import load_model_safely, ClassificationThread
 
+# Import file selection widgets
+from file_widgets import FolderSelectionWidget, FileSelectionWidget
 
-class LicenseDialog(QDialog):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.setWindowTitle("Dotmini MCX License Verification")
-        self.setFixedSize(450, 280)
-        self.setWindowFlag(Qt.WindowType.FramelessWindowHint)
-        
-        # Create license manager
-        self.license_manager = LicenseManager()
-        
-        # Main layout
-        layout = QVBoxLayout()
-        layout.setSpacing(15)
-        
-        # Content widget for styling
-        content_widget = QWidget()
-        content_widget.setObjectName("splashContent")
-        content_layout = QVBoxLayout(content_widget)
-        content_layout.setContentsMargins(25, 25, 25, 25)
-        content_layout.setSpacing(15)
-        
-        # Logo and title
-        title_label = QLabel(f"Dotmini MCX v{APP_VERSION}")
-        title_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        font = title_label.font()
-        font.setPointSize(18)
-        font.setBold(True)
-        title_label.setFont(font)
-        
-        logo_label = QLabel("🔍")
-        logo_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        logo_label.setStyleSheet("font-size: 48px;")
-        
-        self.msg_label = QLabel("Please enter your license key to continue:")
-        
-        # Add a card for the license input
-        license_frame = QFrame()
-        license_frame.setObjectName("card")
-        license_layout = QVBoxLayout(license_frame)
-        license_layout.setContentsMargins(15, 15, 15, 15)
-        
-        self.key_input = QLineEdit()
-        self.key_input.setPlaceholderText("Enter your license key...")
-        
-        # Create keyboard shortcuts for paste
-        paste_shortcut = QAction("Paste", self)
-        paste_shortcut.setShortcut(QKeySequence.StandardKey.Paste)
-        paste_shortcut.triggered.connect(self.paste_from_clipboard)
-        self.key_input.addAction(paste_shortcut)
-        
-        # Remember license checkbox
-        self.remember_checkbox = QCheckBox("Remember license key")
-        self.remember_checkbox.setChecked(True)
-        
-        license_layout.addWidget(self.key_input)
-        license_layout.addWidget(self.remember_checkbox)
-        
-        # Buttons
-        button_layout = QHBoxLayout()
-        
-        self.verify_button = QPushButton("Verify License")
-        self.verify_button.clicked.connect(self.verify_license)
-        
-        quit_button = QPushButton("Quit")
-        quit_button.setObjectName("secondaryButton")
-        quit_button.clicked.connect(self.reject)
-        
-        button_layout.addWidget(quit_button)
-        button_layout.addWidget(self.verify_button)
-        
-        # Add widgets to layout
-        content_layout.addWidget(logo_label)
-        content_layout.addWidget(title_label)
-        content_layout.addWidget(self.msg_label)
-        content_layout.addWidget(license_frame)
-        content_layout.addLayout(button_layout)
-        
-        # Add shadow effect to content widget
-        shadow = QGraphicsDropShadowEffect(self)
-        shadow.setBlurRadius(20)
-        shadow.setColor(QColor(0, 0, 0, 60))
-        shadow.setOffset(0, 2)
-        content_widget.setGraphicsEffect(shadow)
-        
-        # Add content widget to main layout
-        layout.addWidget(content_widget)
-        self.setLayout(layout)
-        
-        # Center on screen
-        self.center_on_screen()
-        
-        # Set default button and key event
-        self.key_input.returnPressed.connect(self.verify_license)
-        self.verify_button.setDefault(True)
-        
-        # Add tip for demo key
-        tip_label = QLabel("Tip: For demo purposes, use key: D1QE80fxUUVcNs4VAAOvNNkJvHHy0dWM")
-        tip_label.setStyleSheet("color: #888888; font-size: 10px;")
-        tip_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        layout.addWidget(tip_label)
+# Import brightness control
+from brightness_control import BrightnessControl
 
-    def center_on_screen(self):
-        screen_geometry = QGuiApplication.primaryScreen().availableGeometry()
-        x = (screen_geometry.width() - self.width()) // 2
-        y = (screen_geometry.height() - self.height()) // 2
-        self.move(x, y)
-    
-    def paste_from_clipboard(self):
-        clipboard = QGuiApplication.clipboard()
-        self.key_input.setText(clipboard.text())
+# Import SettingsDialog
+from settings_dialog import SettingsDialog
 
-    def verify_license(self):
-        entered_key = self.key_input.text().strip()
-        
-        if self.license_manager.correct_key == entered_key:
-            if self.remember_checkbox.isChecked():
-                self.license_manager.save_license(entered_key)
-            self.accept()
-        else:
-            self.msg_label.setText("Invalid license key. Please try again.")
-            self.msg_label.setStyleSheet("color: #FF5555;")
-            # Shake animation to indicate error
-            self.shake_animation()
-    
-    def shake_animation(self):
-        """Create a shake animation to indicate invalid input"""
-        self.animation = QPropertyAnimation(self, b"pos")
-        self.animation.setDuration(100)
-        
-        current_pos = self.pos()
-        x, y = current_pos.x(), current_pos.y()
-        
-        self.animation.setKeyValueAt(0, QPoint(x, y))
-        self.animation.setKeyValueAt(0.1, QPoint(x + 10, y))
-        self.animation.setKeyValueAt(0.2, QPoint(x - 10, y))
-        self.animation.setKeyValueAt(0.3, QPoint(x + 10, y))
-        self.animation.setKeyValueAt(0.4, QPoint(x - 10, y))
-        self.animation.setKeyValueAt(0.5, QPoint(x + 10, y))
-        self.animation.setKeyValueAt(0.6, QPoint(x - 10, y))
-        self.animation.setKeyValueAt(0.7, QPoint(x + 10, y))
-        self.animation.setKeyValueAt(0.8, QPoint(x - 10, y))
-        self.animation.setKeyValueAt(0.9, QPoint(x + 5, y))
-        self.animation.setKeyValueAt(1, QPoint(x, y))
-        
-        self.animation.start()
-
-
-# Custom model loader to fix DepthwiseConv2D issue
-def load_model_safely(model_path):
-    """Load model with custom objects to handle the DepthwiseConv2D issue"""
-    try:
-        # First try to load the model with standard TF method
-        model = tf.keras.models.load_model(model_path, compile=False)
-        return model
-    except Exception as e:
-        error_str = str(e)
-        
-        # Check if it's the specific DepthwiseConv2D error we're trying to fix
-        if "DepthwiseConv2D" in error_str and "groups" in error_str:
-            print("Detected DepthwiseConv2D issue, trying custom loader...")
-            
-            # Define a custom DepthwiseConv2D layer to handle the 'groups' parameter
-            class CustomDepthwiseConv2D(tf.keras.layers.DepthwiseConv2D):
-                def __init__(self, **kwargs):
-                    # Remove 'groups' parameter if present
-                    if 'groups' in kwargs:
-                        del kwargs['groups']
-                    super().__init__(**kwargs)
-            
-            # Load the model with the custom object
-            model = tf.keras.models.load_model(
-                model_path,
-                custom_objects={'DepthwiseConv2D': CustomDepthwiseConv2D},
-                compile=False
-            )
-            return model
-        else:
-            # If it's a different error, re-raise
-            raise
-
-
-class ClassificationThread(QThread):
-    progress_update = pyqtSignal(int, int)
-    result_update = pyqtSignal(str, str, str)
-    finished = pyqtSignal()
-    error = pyqtSignal(str)
-
-    def __init__(self, model_path, label_path, input_folders, output_folder, batch_size=16):
-        super().__init__()
-        self.model_path = model_path
-        self.label_path = label_path
-        self.input_folders = input_folders
-        self.output_folder = output_folder
-        self.batch_size = batch_size
-        self.is_running = True
-        
-        # Performance optimization
-        self.max_workers = min(os.cpu_count() or 4, 8)  # Limit thread count
-
-    def run(self):
-        try:
-            # Load the model using our safe loader
-            model = load_model_safely(self.model_path)
-            
-            # Load labels
-            with open(self.label_path, 'r') as f:
-                labels = [line.strip() for line in f.readlines()]
-            
-            # Get all image files from input folders
-            image_files = []
-            for folder in self.input_folders:
-                for ext in ['*.jpg', '*.jpeg', '*.png', '*.bmp']:
-                    image_files.extend(glob.glob(os.path.join(folder, ext)))
-                    image_files.extend(glob.glob(os.path.join(folder, '**', ext), recursive=True))
-            
-            # Remove duplicates
-            image_files = list(set(image_files))
-            total_files = len(image_files)
-            processed = 0
-            
-            # Create output folders for all classes in advance
-            for class_name in labels:
-                class_folder = os.path.join(self.output_folder, class_name)
-                os.makedirs(class_folder, exist_ok=True)
-            
-            # Process images in batches for better performance
-            for i in range(0, total_files, self.batch_size):
-                if not self.is_running:
-                    break
-                
-                batch_files = image_files[i:i+self.batch_size]
-                batch_results = []
-                
-                # Process batch in parallel
-                with concurrent.futures.ThreadPoolExecutor(max_workers=self.max_workers) as executor:
-                    futures = []
-                    for img_path in batch_files:
-                        futures.append(executor.submit(self.process_image, img_path, model, labels))
-                    
-                    for future in concurrent.futures.as_completed(futures):
-                        if not self.is_running:
-                            break
-                        
-                        result = future.result()
-                        if result:
-                            batch_results.append(result)
-                            processed += 1
-                            self.progress_update.emit(processed, total_files)
-                
-                # Update UI with batch results
-                for result in batch_results:
-                    if not self.is_running:
-                        break
-                    
-                    img_path, predicted_class, confidence, output_path = result
-                    self.result_update.emit(img_path, predicted_class, f"{confidence:.2f}")
-            
-            self.finished.emit()
-            
-        except Exception as e:
-            self.error.emit(f"Error in classification process: {str(e)}")
-            self.finished.emit()
-    
-    def process_image(self, img_path, model, labels):
-        """Process a single image and return the results"""
-        try:
-            # Preprocess image
-            img = Image.open(img_path).convert('RGB')
-            img = img.resize((224, 224))  # Adjust size according to your model
-            img_array = np.array(img) / 255.0  # Normalize
-            img_array = np.expand_dims(img_array, axis=0)
-            
-            # Make prediction
-            predictions = model.predict(img_array, verbose=0)  # Disable verbose output
-            predicted_class_idx = np.argmax(predictions[0])
-            predicted_class = labels[predicted_class_idx]
-            confidence = float(predictions[0][predicted_class_idx])
-            
-            # Prepare output path
-            class_folder = os.path.join(self.output_folder, predicted_class)
-            filename = os.path.basename(img_path)
-            output_path = os.path.join(class_folder, filename)
-            
-            # Save the image to the output folder
-            img.save(output_path)
-            
-            return (img_path, predicted_class, confidence, output_path)
-        
-        except Exception as e:
-            self.error.emit(f"Error processing {img_path}: {str(e)}")
-            return None
-    
-    def stop(self):
-        self.is_running = False
-
-
-class FolderSelectionWidget(QWidget):
-    def __init__(self, title, multiple=False, placeholder_text="Select folder..."):
-        super().__init__()
-        self.multiple = multiple
-        
-        layout = QVBoxLayout()
-        layout.setContentsMargins(0, 0, 0, 0)
-        
-        title_label = QLabel(title)
-        title_label.setStyleSheet("font-weight: bold; font-size: 14px;")
-        
-        self.path_input = QLineEdit()
-        self.path_input.setPlaceholderText(placeholder_text)
-        self.path_input.setReadOnly(True)
-        
-        browse_button = QPushButton("Browse")
-        browse_button.setFixedWidth(100)
-        browse_button.clicked.connect(self.browse_folder)
-        
-        input_layout = QHBoxLayout()
-        input_layout.addWidget(self.path_input)
-        input_layout.addWidget(browse_button)
-        
-        if multiple:
-            self.folder_list = QListWidget()
-            self.folder_list.setMaximumHeight(120)
-            self.folder_list.setDragDropMode(QListWidget.DragDropMode.DragDrop)
-            self.folder_list.setAcceptDrops(True)
-            self.folder_list.viewport().setAcceptDrops(True)
-            self.folder_list.setDropIndicatorShown(True)
-            self.folder_list.dragEnterEvent = self.dragEnterEvent
-            self.folder_list.dragMoveEvent = self.dragMoveEvent
-            self.folder_list.dropEvent = self.dropEvent
-
-            # Add a drop zone label
-            self.drop_label = QLabel("Drop folders here or use Browse button")
-            self.drop_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-            self.drop_label.setStyleSheet("color: #888888; font-style: italic; padding: 10px;")
-            
-            layout.addWidget(title_label)
-            layout.addLayout(input_layout)
-            layout.addWidget(self.drop_label)
-            layout.addWidget(self.folder_list)
-            
-            buttons_layout = QHBoxLayout()
-            
-            remove_button = QPushButton("Remove Selected")
-            remove_button.setObjectName("secondaryButton")
-            remove_button.clicked.connect(self.remove_selected)
-            
-            clear_button = QPushButton("Clear All")
-            clear_button.setObjectName("secondaryButton")
-            clear_button.clicked.connect(self.clear_all)
-            
-            buttons_layout.addWidget(remove_button)
-            buttons_layout.addWidget(clear_button)
-            
-            layout.addLayout(buttons_layout)
-        else:
-            # Setup drop area for single folder selection too
-            self.setAcceptDrops(True)
-            
-            # Add a visual drop zone
-            drop_zone = QFrame()
-            drop_zone.setObjectName("card")
-            drop_zone.setMinimumHeight(80)
-            drop_layout = QVBoxLayout(drop_zone)
-            
-            self.drop_label = QLabel("Drop folder here or use Browse button")
-            self.drop_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-            self.drop_label.setStyleSheet("color: #888888; font-style: italic;")
-            
-            drop_layout.addWidget(self.drop_label)
-            
-            layout.addWidget(title_label)
-            layout.addLayout(input_layout)
-            layout.addWidget(drop_zone)
-        
-        self.setLayout(layout)
-        
-    def browse_folder(self):
-        if self.multiple:
-            folder = QFileDialog.getExistingDirectory(self, "Select Folder")
-            if folder:
-                if self.folder_list.findItems(folder, Qt.MatchFlag.MatchExactly) == []:
-                    self.folder_list.addItem(folder)
-        else:
-            folder = QFileDialog.getExistingDirectory(self, "Select Folder")
-            if folder:
-                self.path_input.setText(folder)
-    
-    def remove_selected(self):
-        for item in self.folder_list.selectedItems():
-            self.folder_list.takeItem(self.folder_list.row(item))
-    
-    def clear_all(self):
-        self.folder_list.clear()
-    
-    def get_paths(self):
-        if self.multiple:
-            return [self.folder_list.item(i).text() for i in range(self.folder_list.count())]
-        else:
-            return self.path_input.text()
-    
-    def set_paths(self, paths):
-        if self.multiple and isinstance(paths, list):
-            self.folder_list.clear()
-            for path in paths:
-                self.folder_list.addItem(path)
-        elif not self.multiple and isinstance(paths, str):
-            self.path_input.setText(paths)
-    
-    def dragEnterEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-    
-    def dragMoveEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-    
-    def dropEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.setDropAction(Qt.DropAction.CopyAction)
-            
-            for url in event.mimeData().urls():
-                path = url.toLocalFile()
-                if os.path.isdir(path):
-                    if self.multiple:
-                        # Check if the path already exists in the list
-                        if self.folder_list.findItems(path, Qt.MatchFlag.MatchExactly) == []:
-                            self.folder_list.addItem(path)
-                    else:
-                        self.path_input.setText(path)
-                        break  # Only use the first folder for single selection
-            
-            event.acceptProposedAction()
-
-
-class FileSelectionWidget(QWidget):
-    def __init__(self, title, file_filter="All Files (*)", placeholder_text="Select file..."):
-        super().__init__()
-        self.file_filter = file_filter
-        self.setAcceptDrops(True)
-        
-        layout = QVBoxLayout()
-        layout.setContentsMargins(0, 0, 0, 0)
-        
-        title_label = QLabel(title)
-        title_label.setStyleSheet("font-weight: bold; font-size: 14px;")
-        
-        self.path_input = QLineEdit()
-        self.path_input.setPlaceholderText(placeholder_text)
-        self.path_input.setReadOnly(True)
-        
-        browse_button = QPushButton("Browse")
-        browse_button.setFixedWidth(100)
-        browse_button.clicked.connect(self.browse_file)
-        
-        input_layout = QHBoxLayout()
-        input_layout.addWidget(self.path_input)
-        input_layout.addWidget(browse_button)
-        
-        # Add a visual drop zone
-        drop_zone = QFrame()
-        drop_zone.setObjectName("card")
-        drop_zone.setMinimumHeight(80)
-        drop_layout = QVBoxLayout(drop_zone)
-        
-        self.drop_label = QLabel("Drop file here or use Browse button")
-        self.drop_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        self.drop_label.setStyleSheet("color: #888888; font-style: italic;")
-        
-        drop_layout.addWidget(self.drop_label)
-        
-        layout.addWidget(title_label)
-        layout.addLayout(input_layout)
-        layout.addWidget(drop_zone)
-        
-        self.setLayout(layout)
-    
-    def browse_file(self):
-        file, _ = QFileDialog.getOpenFileName(self, "Select File", "", self.file_filter)
-        if file:
-            self.path_input.setText(file)
-    
-    def get_path(self):
-        return self.path_input.text()
-    
-    def set_path(self, path):
-        self.path_input.setText(path)
-    
-    def dragEnterEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-    
-    def dragMoveEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.acceptProposedAction()
-    
-    def dropEvent(self, event):
-        if event.mimeData().hasUrls():
-            event.setDropAction(Qt.DropAction.CopyAction)
-            
-            for url in event.mimeData().urls():
-                path = url.toLocalFile()
-                if os.path.isfile(path):
-                    # Check if file matches the filter
-                    file_ext = os.path.splitext(path)[1].lower()
-                    
-                    # Extract extensions from filter
-                    filter_exts = []
-                    for filter_part in self.file_filter.split(';;'):
-                        if '(*)' in filter_part:  # All files
-                            self.path_input.setText(path)
-                            event.acceptProposedAction()
-                            return
-                        
-                        ext_part = filter_part.split('(')[1].split(')')[0] if '(' in filter_part else ""
-                        exts = [e.strip().lower() for e in ext_part.split() if e.startswith('*.')]
-                        filter_exts.extend([e[1:] for e in exts])  # Remove the '*'
-                    
-                    # If no specific extensions or file matches the filter
-                    if not filter_exts or file_ext in filter_exts:
-                        self.path_input.setText(path)
-                        break
-            
-            event.acceptProposedAction()
-
-
-class BrightnessControl(QWidget):
-    brightnessChanged = pyqtSignal(int)
-    
-    def __init__(self, initial_value=100):
-        super().__init__()
-        
-        layout = QVBoxLayout()
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(5)
-        
-        # Header
-        header_layout = QHBoxLayout()
-        
-        brightness_label = QLabel("Brightness")
-        brightness_label.setStyleSheet("font-weight: bold; font-size: 14px;")
-        
-        self.value_label = QLabel(f"{initial_value}%")
-        
-        header_layout.addWidget(brightness_label)
-        header_layout.addStretch()
-        header_layout.addWidget(self.value_label)
-        
-        # Slider
-        self.slider = QSlider(Qt.Orientation.Horizontal)
-        self.slider.setMinimum(50)  # Minimum 50% brightness
-        self.slider.setMaximum(150)  # Maximum 150% brightness
-        self.slider.setValue(initial_value)
-        self.slider.setTracking(True)
-        
-        self.slider.valueChanged.connect(self.on_value_changed)
-        
-        # Controls
-        controls_layout = QHBoxLayout()
-        
-        dim_button = QPushButton("Dim")
-        dim_button.setObjectName("secondaryButton")
-        dim_button.setFixedWidth(80)
-        dim_button.clicked.connect(self.on_dim_clicked)
-        
-        reset_button = QPushButton("Reset")
-        reset_button.setObjectName("secondaryButton")
-        reset_button.setFixedWidth(80)
-        reset_button.clicked.connect(self.on_reset_clicked)
-        
-        bright_button = QPushButton("Bright")
-        bright_button.setObjectName("secondaryButton")
-        bright_button.setFixedWidth(80)
-        bright_button.clicked.connect(self.on_bright_clicked)
-        
-        controls_layout.addWidget(dim_button)
-        controls_layout.addStretch()
-        controls_layout.addWidget(reset_button)
-        controls_layout.addStretch()
-        controls_layout.addWidget(bright_button)
-        
-        # Add all to main layout
-        layout.addLayout(header_layout)
-        layout.addWidget(self.slider)
-        layout.addLayout(controls_layout)
-        
-        self.setLayout(layout)
-    
-    def on_value_changed(self, value):
-        self.value_label.setText(f"{value}%")
-        self.brightnessChanged.emit(value)
-    
-    def on_dim_clicked(self):
-        current_value = self.slider.value()
-        new_value = max(current_value - 10, self.slider.minimum())
-        self.slider.setValue(new_value)
-    
-    def on_reset_clicked(self):
-        self.slider.setValue(100)
-    
-    def on_bright_clicked(self):
-        current_value = self.slider.value()
-        new_value = min(current_value + 10, self.slider.maximum())
-        self.slider.setValue(new_value)
-    
-    def get_brightness(self):
-        return self.slider.value()
-    
-    def set_brightness(self, value):
-        self.slider.setValue(value)
-
-
-class SettingsDialog(QDialog):
-    settingsChanged = pyqtSignal()
-    
-    def __init__(self, theme_manager, parent=None):
-        super().__init__(parent)
-        self.theme_manager = theme_manager
-        
-        self.setWindowTitle("Dotmini MCX Settings")
-        self.setMinimumWidth(450)
-        
-        layout = QVBoxLayout()
-        layout.setSpacing(20)
-        
-        # Theme selection
-        theme_frame = QFrame()
-        theme_frame.setObjectName("card")
-        theme_layout = QVBoxLayout(theme_frame)
-        
-        theme_label = QLabel("Theme")
-        theme_label.setStyleSheet("font-weight: bold; font-size: 16px;")
-        
-        self.theme_auto = QCheckBox("Auto (use system theme)")
-        self.theme_auto.setChecked(True)
-        
-        theme_button_layout = QHBoxLayout()
-        
-        self.light_theme_button = QPushButton("Light")
-        self.light_theme_button.setCheckable(True)
-        
-        self.dark_theme_button = QPushButton("Dark")
-        self.dark_theme_button.setCheckable(True)
-        
-        # Set initial state
-        if self.theme_manager.current_theme == "light":
-            self.theme_auto.setChecked(False)
-            self.light_theme_button.setChecked(True)
-            self.dark_theme_button.setChecked(False)
-        elif self.theme_manager.current_theme == "dark":
-            self.theme_auto.setChecked(False)
-            self.light_theme_button.setChecked(False)
-            self.dark_theme_button.setChecked(True)
-        
-        # Connect theme change signals
-        self.theme_auto.toggled.connect(self.on_theme_auto_changed)
-        self.light_theme_button.clicked.connect(self.on_light_theme_clicked)
-        self.dark_theme_button.clicked.connect(self.on_dark_theme_clicked)
-        
-        theme_button_layout.addWidget(self.light_theme_button)
-        theme_button_layout.addWidget(self.dark_theme_button)
-        
-        theme_layout.addWidget(theme_label)
-        theme_layout.addWidget(self.theme_auto)
-        theme_layout.addLayout(theme_button_layout)
-        
-        # Brightness control
-        brightness_frame = QFrame()
-        brightness_frame.setObjectName("card")
-        brightness_layout = QVBoxLayout(brightness_frame)
-        
-        brightness_label = QLabel("Appearance")
-        brightness_label.setStyleSheet("font-weight: bold; font-size: 16px;")
-        
-        self.brightness_control = BrightnessControl(self.theme_manager.brightness)
-        self.brightness_control.brightnessChanged.connect(self.on_brightness_changed)
-        
-        brightness_layout.addWidget(brightness_label)
-        brightness_layout.addWidget(self.brightness_control)
-        
-        # Classification settings
-        classification_frame = QFrame()
-        classification_frame.setObjectName("card")
-        classification_layout = QVBoxLayout(classification_frame)
-        
-        classification_label = QLabel("Classification Settings")
-        classification_label.setStyleSheet("font-weight: bold; font-size: 16px;")
-        
-        self.batch_size_label = QLabel("Batch Size:")
-        self.batch_size_combo = QComboBox()
-        for size in [1, 4, 8, 16, 32, 64]:
-            self.batch_size_combo.addItem(str(size))
-        self.batch_size_combo.setCurrentText("16")  # Default batch size
-        
-        batch_layout = QHBoxLayout()
-        batch_layout.addWidget(self.batch_size_label)
-        batch_layout.addWidget(self.batch_size_combo)
-        
-        classification_layout.addWidget(classification_label)
-        classification_layout.addLayout(batch_layout)
-        
-        # Buttons
-        button_layout = QHBoxLayout()
-        
-        cancel_button = QPushButton("Cancel")
-        cancel_button.setObjectName("secondaryButton")
-        cancel_button.clicked.connect(self.reject)
-        
-        save_button = QPushButton("Save Settings")
-        save_button.clicked.connect(self.accept)
-        
-        button_layout.addWidget(cancel_button)
-        button_layout.addWidget(save_button)
-        
-        # Add all to main layout
-        layout.addWidget(theme_frame)
-        layout.addWidget(brightness_frame)
-        layout.addWidget(classification_frame)
-        layout.addLayout(button_layout)
-        
-        self.setLayout(layout)
-    
-    def on_theme_auto_changed(self, checked):
-        if checked:
-            # Enable system theme detection
-            if DARKDETECT_AVAILABLE:
-                self.theme_manager.detect_system_theme()
-                self.theme_manager.apply_theme()
-            
-            # Update buttons
-            self.light_theme_button.setChecked(False)
-            self.dark_theme_button.setChecked(False)
-    
-    def on_light_theme_clicked(self):
-        self.theme_auto.setChecked(False)
-        self.light_theme_button.setChecked(True)
-        self.dark_theme_button.setChecked(False)
-        self.theme_manager.current_theme = "light"
-        self.theme_manager.apply_theme()
-    
-    def on_dark_theme_clicked(self):
-        self.theme_auto.setChecked(False)
-        self.light_theme_button.setChecked(False)
-        self.dark_theme_button.setChecked(True)
-        self.theme_manager.current_theme = "dark"
-        self.theme_manager.apply_theme()
-    
-    def on_brightness_changed(self, value):
-        self.theme_manager.set_brightness(value)
-    
-    def save_settings(self):
-        # Save settings
-        self.theme_manager.save_settings()
-        
-        # Emit signal to notify main window
-        self.settingsChanged.emit()
-
+# Need DARKDETECT_AVAILABLE for SettingsDialog, ensure it's accessible or pass it if it's defined in theme_manager
+# For now, assuming SettingsDialog correctly imports it or handles its absence.
+# from theme_manager import DARKDETECT_AVAILABLE # This is already imported by settings_dialog.py
 
 class DotminiMCX(QMainWindow):
+    """The main window class for the Dotmini MCX application.
+    
+    This class sets up the user interface, manages application state,
+    and handles interactions for image classification tasks.
+    """
     def __init__(self):
+        """Initializes the main application window and its components."""
         super().__init__()
         
         # Initialize settings
         self.settings = QSettings(ORGANIZATION_NAME, APP_NAME)
         
         # Create theme manager
-        self.theme_manager = None  # Will be set after QApplication is created
+        # self.theme_manager will be set by the main() function after QApplication is created.
+        self.theme_manager = None 
         
         # Initialize UI
         self.init_ui()
         
         # Member variables
         self.classification_thread = None
-        self.batch_size = 16
-        self.recent_paths = []
+        self.batch_size = 16 # Default batch size, loaded from settings in load_settings()
+        self.recent_paths = [] # For storing recent file/folder paths
     
     def init_ui(self):
+        """Sets up the main window's user interface layout and widgets."""
         self.setWindowTitle("Dotmini MCX - Image Classification Tool")
         self.setMinimumSize(900, 700)
         
@@ -1511,12 +293,13 @@ class DotminiMCX(QMainWindow):
         self.center_on_screen()
     
     def create_toolbar(self):
+        """Creates the main application toolbar with theme and settings actions."""
         toolbar = QToolBar("Main Toolbar")
         toolbar.setMovable(False)
         
         # Add theme toggle button to toolbar
-        theme_icon = "🌙" if self.theme_manager and self.theme_manager.current_theme == "light" else "☀️"
-        theme_action = QAction(f"{theme_icon} Toggle Theme", self)
+        # The icon will be updated by update_theme_ui
+        theme_action = QAction("Toggle Theme", self) 
         theme_action.triggered.connect(self.toggle_theme)
         theme_action.setToolTip("Switch between light and dark mode")
         toolbar.addAction(theme_action)
@@ -1535,6 +318,7 @@ class DotminiMCX(QMainWindow):
         self.addToolBar(toolbar)
     
     def create_menu_bar(self):
+        """Creates the main application menu bar."""
         menubar = self.menuBar()
         
         # File Menu
@@ -1584,26 +368,24 @@ class DotminiMCX(QMainWindow):
         help_menu.addAction(docs_action)
     
     def toggle_theme(self):
+        """Toggles the application theme between light and dark modes."""
         if self.theme_manager:
             self.theme_manager.toggle_theme()
-            # Update toolbar button
-            for action in self.findChild(QToolBar).actions():
-                if "Toggle Theme" in action.text():
-                    theme_icon = "🌙" if self.theme_manager.current_theme == "light" else "☀️"
+            # Update toolbar button (update_theme_ui is called by ThemeManager.apply_theme)
+    
+    def update_theme_ui(self):
+        """Updates specific UI elements when the theme changes (e.g., toolbar icons)."""
+        # Update toolbar button
+        toolbar = self.findChild(QToolBar)
+        if toolbar:
+            for action in toolbar.actions():
+                if "Toggle Theme" in action.text() or "🌙" in action.text() or "☀️" in action.text(): # Check for text or icons
+                    theme_icon = "🌙" if self.theme_manager and self.theme_manager.current_theme == "light" else "☀️"
                     action.setText(f"{theme_icon} Toggle Theme")
                     break
     
-    def update_theme_ui(self):
-        """Called when theme is changed to update UI elements that need specific handling"""
-        # Update toolbar button
-        for action in self.findChild(QToolBar).actions():
-            if "Toggle Theme" in action.text():
-                theme_icon = "🌙" if self.theme_manager.current_theme == "light" else "☀️"
-                action.setText(f"{theme_icon} Toggle Theme")
-                break
-    
     def add_shadow(self, widget):
-        """Add shadow effect to a widget"""
+        """Adds a pre-configured drop shadow effect to the given widget."""
         shadow = QGraphicsDropShadowEffect(self)
         shadow.setBlurRadius(15)
         shadow.setColor(QColor(0, 0, 0, 40))
@@ -1611,10 +393,12 @@ class DotminiMCX(QMainWindow):
         widget.setGraphicsEffect(shadow)
     
     def update_recent_menu(self):
-        """Update the recent files menu with saved recent paths"""
+        """Populates the 'Recent Files' menu with recently used paths."""
         self.recent_menu.clear()
         
         recent_paths = self.settings.value("recent/paths", [])
+        if not isinstance(recent_paths, list): recent_paths = [] # Ensure it's a list
+
         if not recent_paths:
             no_recent = QAction("No recent files", self)
             no_recent.setEnabled(False)
@@ -1622,7 +406,8 @@ class DotminiMCX(QMainWindow):
             return
         
         for path in recent_paths:
-            action = QAction(os.path.basename(path), self)
+            if not path: continue # Skip empty paths
+            action = QAction(os.path.basename(path) or path, self) # Use full path if basename is empty
             action.setData(path)
             action.setToolTip(path)
             action.triggered.connect(self.load_recent_path)
@@ -1634,7 +419,7 @@ class DotminiMCX(QMainWindow):
         self.recent_menu.addAction(clear_action)
     
     def load_recent_path(self):
-        """Load a selected recent path"""
+        """Loads a path selected from the 'Recent Files' menu into the appropriate widget."""
         action = self.sender()
         if action:
             path = action.data()
@@ -1646,19 +431,20 @@ class DotminiMCX(QMainWindow):
                 elif ext in ['.txt']:
                     self.label_file_widget.set_path(path)
             elif os.path.isdir(path):
-                # For directories, try to determine the context
-                if self.input_folders_widget.get_paths():
-                    self.output_folder_widget.set_paths(path)
-                else:
+                # For directories, assume it's for input or output
+                # A more sophisticated approach might be needed if context is ambiguous
+                if not self.input_folders_widget.get_paths(): # If input is empty, populate it
                     self.input_folders_widget.set_paths([path])
+                else: # Otherwise, assume it's for output
+                    self.output_folder_widget.set_paths(path)
     
     def add_to_recent_files(self, path):
-        """Add a path to recent files"""
+        """Adds a given path to the list of recent files in settings."""
         if not path:
             return
             
         recent_paths = self.settings.value("recent/paths", [])
-        if not isinstance(recent_paths, list):
+        if not isinstance(recent_paths, list): # Ensure it's a list
             recent_paths = []
             
         # Remove duplicates
@@ -1672,20 +458,28 @@ class DotminiMCX(QMainWindow):
         recent_paths = recent_paths[:10]
         
         self.settings.setValue("recent/paths", recent_paths)
-        self.update_recent_menu()
+        self.update_recent_menu() # Refresh the menu
     
     def clear_recent_files(self):
-        """Clear all recent files"""
+        """Clears all entries from the 'Recent Files' list and menu."""
         self.settings.setValue("recent/paths", [])
         self.update_recent_menu()
     
     def show_settings(self):
+        """Displays the application settings dialog."""
+        if not self.theme_manager: return # Guard against missing theme_manager
         settings_dialog = SettingsDialog(self.theme_manager, self)
         if settings_dialog.exec():
-            settings_dialog.save_settings()
-            self.load_settings()
+            settings_dialog.save_settings() # Dialog now handles saving its own relevant settings
+            # Batch size is saved by dialog, now main window needs to update its own batch_size
+            self.batch_size = settings_dialog.get_batch_size() 
+            self.settings.setValue("classification/batch_size", self.batch_size) # Also save to QSettings
+
+            if self.theme_manager:
+                 self.theme_manager.apply_theme() # Re-apply theme if settings changed it
     
     def show_about(self):
+        """Displays the 'About' dialog with application information."""
         about_text = f"""<h2>Dotmini MCX v{APP_VERSION}</h2>
 <p>A machine learning image classification tool with modern UI.</p>
 <p>© 2023 DotminiTech</p>"""
@@ -1693,7 +487,7 @@ class DotminiMCX(QMainWindow):
         QMessageBox.about(self, "About Dotmini MCX", about_text)
     
     def show_documentation(self):
-        """Show documentation or help about the app"""
+        """Displays a simple documentation/help message box."""
         help_text = """
 <h2>Dotmini MCX - Quick Guide</h2>
 
@@ -1729,83 +523,75 @@ class DotminiMCX(QMainWindow):
         msg_box.exec()
     
     def load_settings(self):
-        """Load application settings"""
+        """Loads application settings from persistent storage (e.g., QSettings)."""
         # Load batch size
-        batch_size = self.settings.value("classification/batch_size", 16, type=int)
-        self.batch_size = batch_size
+        self.batch_size = self.settings.value("classification/batch_size", 16, type=int)
         
         # Load last used paths
         last_model_path = self.settings.value("paths/model", "")
-        if last_model_path:
-            self.model_file_widget.set_path(last_model_path)
+        if last_model_path: self.model_file_widget.set_path(last_model_path)
         
         last_label_path = self.settings.value("paths/labels", "")
-        if last_label_path:
-            self.label_file_widget.set_path(last_label_path)
+        if last_label_path: self.label_file_widget.set_path(last_label_path)
         
         last_output_path = self.settings.value("paths/output", "")
-        if last_output_path:
-            self.output_folder_widget.set_paths(last_output_path)
+        if last_output_path: self.output_folder_widget.set_paths(last_output_path)
         
         # Load input folders
         input_folders = self.settings.value("paths/input_folders", [])
-        if input_folders:
+        if isinstance(input_folders, list) and input_folders: # Ensure it's a non-empty list
             self.input_folders_widget.set_paths(input_folders)
         
         # Update recent files menu
         self.update_recent_menu()
     
     def save_settings(self):
-        """Save application settings"""
+        """Saves current application settings to persistent storage."""
         # Save batch size
         self.settings.setValue("classification/batch_size", self.batch_size)
         
         # Save paths
         model_path = self.model_file_widget.get_path()
         label_path = self.label_file_widget.get_path()
-        output_path = self.output_folder_widget.get_paths()
-        input_folders = self.input_folders_widget.get_paths()
+        output_path = self.output_folder_widget.get_paths() # This is a string
+        input_folders = self.input_folders_widget.get_paths() # This is a list of strings
         
-        self.settings.setValue("paths/model", model_path)
-        self.settings.setValue("paths/labels", label_path)
-        self.settings.setValue("paths/output", output_path)
-        self.settings.setValue("paths/input_folders", input_folders)
+        if model_path: self.settings.setValue("paths/model", model_path)
+        if label_path: self.settings.setValue("paths/labels", label_path)
+        if output_path: self.settings.setValue("paths/output", output_path) # Save single path string
+        if input_folders: self.settings.setValue("paths/input_folders", input_folders) # Save list
         
-        # Add to recent files
-        if model_path:
-            self.add_to_recent_files(model_path)
-        if label_path:
-            self.add_to_recent_files(label_path)
-        if output_path:
-            self.add_to_recent_files(output_path)
+        # Add to recent files (these methods also update the menu)
+        if model_path: self.add_to_recent_files(model_path)
+        if label_path: self.add_to_recent_files(label_path)
+        if output_path: self.add_to_recent_files(output_path)
         for folder in input_folders:
-            self.add_to_recent_files(folder)
+            if folder: self.add_to_recent_files(folder)
     
     def center_on_screen(self):
+        """Centers the main window on the primary screen."""
         screen_geometry = QGuiApplication.primaryScreen().availableGeometry()
         x = (screen_geometry.width() - self.width()) // 2
         y = (screen_geometry.height() - self.height()) // 2
         self.move(x, y)
     
     def start_classification(self):
+        """Starts the image classification process in a separate thread."""
         # Validate inputs
         input_folders = self.input_folders_widget.get_paths()
         model_path = self.model_file_widget.get_path()
         label_path = self.label_file_widget.get_path()
-        output_folder = self.output_folder_widget.get_paths()
+        output_folder = self.output_folder_widget.get_paths() # Single string
         
         if not input_folders:
             QMessageBox.warning(self, "Input Error", "Please select at least one input folder.")
             return
-            
         if not model_path:
             QMessageBox.warning(self, "Input Error", "Please select a model file.")
             return
-            
         if not label_path:
             QMessageBox.warning(self, "Input Error", "Please select a labels file.")
             return
-            
         if not output_folder:
             QMessageBox.warning(self, "Input Error", "Please select an output folder.")
             return
@@ -1844,30 +630,30 @@ class DotminiMCX(QMainWindow):
         self.classification_thread.start()
     
     def cancel_classification(self):
+        """Cancels the ongoing classification process."""
         if self.classification_thread is not None and self.classification_thread.isRunning():
             self.classification_thread.stop()
-            self.classification_thread.wait()
-            self.classification_finished()
-            self.add_result("", "Classification canceled by user", "")
-            self.statusBar().showMessage("Classification canceled.")
+            self.classification_thread.wait() # Wait for thread to finish cleanly
+            # classification_finished() will be called by the thread's finished signal
+            self.statusBar().showMessage("Classification canceled by user.")
+            self.add_result("", "Classification canceled by user.", "") # Add a message to results
     
     def update_progress(self, value, maximum):
+        """Updates the progress bar and status message during classification."""
         self.progress_bar.setMaximum(maximum)
         self.progress_bar.setValue(value)
         self.statusBar().showMessage(f"Processing images: {value}/{maximum} ({int(value/maximum*100)}%)")
     
     def add_result(self, file_path, class_name, confidence):
-        if file_path:
-            # Create a custom widget item for better display
+        """Adds a classification result to the results list and updates statistics."""
+        if file_path: # Actual result item
             item = QListWidgetItem()
-            
-            # Try to create a thumbnail for the image
             try:
-                pixmap = QPixmap(file_path).scaled(40, 40, Qt.AspectRatioMode.KeepAspectRatio)
+                pixmap = QPixmap(file_path).scaled(40, 40, Qt.AspectRatioMode.KeepAspectRatio, Qt.TransformationMode.SmoothTransformation)
                 item.setIcon(QIcon(pixmap))
-            except:
-                # If thumbnail creation fails, use a generic icon
-                item.setIcon(QIcon())
+            except Exception as e: # Catch potential errors with QPixmap
+                print(f"Error creating thumbnail for {file_path}: {e}")
+                item.setIcon(QIcon()) # Use a generic icon
             
             item_text = f"{os.path.basename(file_path)} → {class_name} (confidence: {confidence})"
             item.setText(item_text)
@@ -1876,24 +662,19 @@ class DotminiMCX(QMainWindow):
             self.results_list.addItem(item)
             
             # Update statistics
-            if class_name in self.class_counts:
-                self.class_counts[class_name] += 1
-            else:
-                self.class_counts[class_name] = 1
-            
+            self.class_counts[class_name] = self.class_counts.get(class_name, 0) + 1
             self.total_processed += 1
             self.update_result_stats()
-        else:
-            # For messages like "canceled"
+        else: # For status messages like "canceled"
             item = QListWidgetItem()
-            item.setText(class_name)
+            item.setText(class_name) # class_name here is the message string
             item.setForeground(QColor("#FF5555"))  # Red color for error/status messages
             self.results_list.addItem(item)
         
         self.results_list.scrollToBottom()
     
     def update_result_stats(self):
-        """Update the results statistics label"""
+        """Updates the label displaying statistics about classification results."""
         if self.total_processed == 0:
             self.results_stats_label.setText("No results yet")
             return
@@ -1911,27 +692,31 @@ class DotminiMCX(QMainWindow):
         self.results_stats_label.setText(stats_text)
     
     def classification_finished(self):
+        """Handles tasks to perform after classification is complete or canceled."""
         self.classify_button.setEnabled(True)
         self.cancel_button.setEnabled(False)
-        self.classification_thread = None
+        self.classification_thread = None # Clear the thread reference
         
-        if self.total_processed > 0:
-            self.statusBar().showMessage(f"Classification complete. Processed {self.total_processed} images.")
-        else:
+        if self.total_processed > 0 and self.statusBar().currentMessage() != "Classification canceled by user.":
+             self.statusBar().showMessage(f"Classification complete. Processed {self.total_processed} images.")
+        elif self.statusBar().currentMessage() != "Classification canceled by user.": # If not canceled and no images processed
             self.statusBar().showMessage("Ready")
     
     def show_error(self, error_message):
+        """Displays an error message in a message box and the status bar."""
         QMessageBox.critical(self, "Error", error_message)
         self.statusBar().showMessage(f"Error: {error_message}")
     
     def clear_results(self):
+        """Clears all classification results from the list and resets statistics."""
         self.results_list.clear()
         self.class_counts = {}
         self.total_processed = 0
         self.update_result_stats()
+        self.statusBar().showMessage("Results cleared.")
     
     def export_results(self):
-        """Export classification results to a CSV file"""
+        """Exports the current classification results to a CSV file."""
         if self.results_list.count() == 0:
             QMessageBox.information(self, "Export Results", "No results to export.")
             return
@@ -1963,32 +748,33 @@ class DotminiMCX(QMainWindow):
                     if len(parts) != 2:
                         continue
                         
-                    filename = parts[0].strip()
+                    file_name_part = parts[0].strip() # Using file_name_part to avoid conflict
                     
                     # Extract class and confidence
                     class_conf = parts[1].strip()
                     class_name = class_conf.split("(confidence:")[0].strip()
                     
                     # Try to extract confidence value
-                    confidence = ""
+                    confidence_val = "" # Using confidence_val to avoid conflict
                     if "confidence:" in class_conf:
-                        confidence = class_conf.split("confidence:")[1].strip()
-                        confidence = confidence.rstrip(")")
+                        confidence_val = class_conf.split("confidence:")[1].strip()
+                        confidence_val = confidence_val.rstrip(")")
                     
                     # Write to CSV
-                    f.write(f'"{filename}","{class_name}",{confidence}\n')
+                    f.write(f'"{file_name_part}","{class_name}",{confidence_val}\n')
                 
             self.statusBar().showMessage(f"Results exported to {filename}")
         except Exception as e:
             QMessageBox.critical(self, "Export Error", f"Error exporting results: {str(e)}")
     
     def closeEvent(self, event):
-        # Save settings before closing
+        """Handles the close event of the main window to save settings."""
         self.save_settings()
         event.accept()
 
 
 def main():
+    """Main function to launch the Dotmini MCX application."""
     # Enable high DPI scaling
     QApplication.setHighDpiScaleFactorRoundingPolicy(
         Qt.HighDpiScaleFactorRoundingPolicy.PassThrough)
@@ -2063,8 +849,9 @@ def main():
     fade_anim.start()
     
     # Progress animation
-    def update_progress():
-        nonlocal progress_value
+    progress_value = 0 # Ensure progress_value is defined in main's scope
+    def update_splash_progress():
+        nonlocal progress_value # Refer to progress_value in main's scope
         progress_value += 1
         splash_progress.setValue(progress_value)
         
@@ -2079,16 +866,15 @@ def main():
             status_label.setText("Almost ready...")
         
         # Re-render widget to pixmap
-        splash_content.render(splash_pixmap)
+        splash_content.render(splash_pixmap) # Re-render with updated progress bar
         splash.setPixmap(splash_pixmap)
         
         if progress_value >= 100:
-            progress_timer.stop()
+            progress_timer.stop() # Stop the timer when progress reaches 100
     
     # Start progress animation
-    progress_value = 0
-    progress_timer = QTimer()
-    progress_timer.timeout.connect(update_progress)
+    progress_timer = QTimer() # progress_timer should also be in main's scope
+    progress_timer.timeout.connect(update_splash_progress)
     progress_timer.start(30)  # Update every 30ms for smooth animation
     
     # Check license first
@@ -2098,18 +884,19 @@ def main():
         license_dialog = LicenseDialog()
         if not license_dialog.exec():
             # Exit if license is not verified or user cancels
-            return
+            return # Exit main()
     
     # Create theme manager
-    theme_manager = ThemeManager(app)
+    theme_manager_instance = ThemeManager(app) # Renamed to avoid conflict
     
     # Create main window
     main_window = DotminiMCX()
-    main_window.theme_manager = theme_manager
-    theme_manager.main_window = main_window
+    main_window.theme_manager = theme_manager_instance # Assign to main_window
+    theme_manager_instance.main_window = main_window # Link back
     
-    # Apply theme based on settings
-    theme_manager.apply_theme()
+    # Apply theme based on settings (ThemeManager constructor already does this)
+    # theme_manager_instance.apply_theme() # Already called in ThemeManager.__init__
+    main_window.update_theme_ui() # Ensure UI elements like toolbar icon are set
     
     # Finish splash screen after a slight delay and fade it out
     def show_main_window():
@@ -2125,7 +912,9 @@ def main():
         # Show main window
         main_window.show()
     
-    QTimer.singleShot(3000, show_main_window)
+    # Ensure splash screen completes its animation before showing main window
+    # Use a timer that considers the progress animation duration
+    QTimer.singleShot(3000, show_main_window) # 30ms * 100 updates = 3000ms
     
     sys.exit(app.exec())
